requirements.txt
torch
torchvision
numpy
pandas>=2.1.0
Pillow>=10.0.0
albumentations>=1.3.0
matplotlib>=3.8.0
seaborn>=0.12.0
grad-cam>=1.4.6
scikit-learn>=1.3.0
tqdm>=4.66.0
requests>=2.31.0
opencv-python